pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "zeststream-voice"
version = "0.4.0"
description = "Brand voice scoring + claim grounding for AI-generated marketing copy"
readme = "src/zeststream_voice/README.md"
license = "MIT"
authors = [{ name = "Joshua Nowak", email = "joshua@zeststream.ai" }]
requires-python = ">=3.11"
dependencies = [
  "pyyaml>=6.0",
  "click>=8.1",
]

[project.optional-dependencies]
embeddings = ["sentence-transformers>=2.7"]
rubric = ["anthropic>=0.34"]
dev = ["pytest>=8.0", "pytest-cov>=4.0"]

[project.scripts]
zeststream-voice = "zeststream_voice.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/zeststream_voice"]
